extract platform-specific helpers from WaveSessionView
Move #if os(macOS) blocks out of WaveSessionView.swift into a
whole-file platform-gated PlatformHelpers.swift file. This satisfies
the multiplatform boundary check that disallows inline #if directives
in shared Concerto views.
